Perfecto! La migración se completó exitosamente. Todos los 500 productos fueron actualizados. Ahora voy a resumir todos los cambios:
---

## ✅ PRECIOS CORREGIDOS A NÚMEROS ENTEROS EN TODO EL SISTEMA

### Cambios realizados:

#### 1. **Modelo de datos** (`/app/backend/models/products.py`)
- `cost_price`: `float` → `int`
- `sale_price`: `float` → `int`
- `last_price`: `float` → `int`
- `ProductCombination.price`: `float` → `int`
- Métodos `tax_amount` y `profit`: ahora retornan `int`

#### 2. **Servicios de sincronización PrestaShop**
- `/app/backend/services/fast_batch_sync_service.py`: convierte precios a entero con `int(round(...))`
- `/app/backend/services/batch_sync_service.py`: convierte precios a entero con `int(round(...))`
- Combinaciones también convertidas a enteros

#### 3. **Frontend - Formularios**
- `/app/frontend/src/components/ProductForm.js`:
  - `parseFloat` → `parseInt` para cost_price y sale_price
  - Inputs con `step="1"` (sin decimales)
  - Display con `toLocaleString('es-CL')` para formato chileno

- `/app/frontend/src/components/SalesForm.js`:
  - Cálculos de total usan `parseInt` en vez de `parseFloat`

#### 4. **Base de datos migrada**
- Script ejecutado: `/app/backend/scripts/migrate_prices_to_int.py`
- **500/500 productos actualizados** ✅
- Todos los `cost_price`, `sale_price`, `last_price` y precios de combinaciones convertidos a enteros

### Resultado:
- ✅ Todos los precios ahora son números enteros (sin decimales)
- ✅ Se muestran con formato chileno: `$12.990` en vez de `$12990.50`
- ✅ Los formularios solo aceptan números enteros
- ✅ Base de datos ya migrada
- ✅ Sincronización PrestaShop convierte automáticamente a enteros

**Backend reiniciado** ✅
**Frontend compilado** ✅

+++ frontend/src/components/SalesForm.js
@@ -205,14 +205,14 @@ const API = `${BACKEND_URL}/api`;
     setShowProductSuggestions(false);
     // Calculate initial total when product is selected
     if (quantity && prod.sale_price) {
-      setTotal((parseFloat(quantity) * prod.sale_price).toString());
+      setTotal((parseInt(quantity) * prod.sale_price).toString());  // Usar enteros
     }
   };
 
   // Update total when quantity changes
   useEffect(() => {
     if (selectedProduct && selectedProduct.sale_price && quantity) {
-      setTotal((parseFloat(quantity) * selectedProduct.sale_price).toString());
+      setTotal((parseInt(quantity) * selectedProduct.sale_price).toString());  // Usar enteros
     } else {
       setTotal('');
     }
